refactor: 토큰 관리 리팩토링
